Move backend manager logic into a mixin
Currently, the network manager code contained everything in one place,
unlike many of the other manageiq providers. Since we are going to be
extending this part with the events, this patch moves most of the logic
targeting towards the backend Nuage services into a separate mixin to
keep the main manager code clean.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>