Introduce Nuage event catcher
This commit introduces the full blown event catcher for Nuage manager
based on ActiveMQ message bus. Further to default authentication, AMQP
is now provided via Qpid Proton gem, along with the corresponding
message handler responsible for the delivery of messages into CFME.
When the worker stream is requested to stop processing messages, the
connection to the AMQP is stopped immediately.

Topics are configurable through the provider's event catcher settings.
Multiple topics can be used allowing users to listen to arbitrary events
and alarms from the provider.

In order to allow connection verification, AMQP connection is actually
started and terminated immediately. This is required to catch differen
kinds of connection errors that may occur.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>