Provide extensive unit tests for legacy refresher
This commit is a preparation to be able to support graph inventory refresh
with the next commit. Basically we provide unit tests for refresh parser so
that we'll be able to assure that the new graph inventory refresh works
identically to the current (legacy) one.

Writing tests we've also discovered that there is a bug in current parser -
the network groups were not associated correctly with the security groups.
Fixed. God bless the unit tests! :D

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>