Implement graph inventory refresh for network manager
With this commit we implement whatever is needed for the
fancy new graph inventory refresh to step in. For now we
keep legacy refresher also in place (switchable in settings.yml),
but the plan is to migrate to the new refresher completely.

Exactly same unit tests that were passing for legacy refresher are
now passing for both the legacy refresher and the new graph refresher.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>