Fix Qpid proton Travis build
Because qpid-proton needs to be built from source, additional system
packages are needed (installed with sudo rights). This caused that the
vendor bundle is created by the root user and resulted in Bundler not
being able to install other gems.

This patch now removes the use of sudo to the bare minimum, namely
installing the packages required to build qpid-proton from source.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>