Add unit tests for VsdClient edge cases
With this commit we add some tests that test two internal
methods of the VsdClient, `get_list` and `get_first`. We only
test those edge cases that are inconvenient to be tested with
VCR cassettes, just to make sure that all the code is run at
least once during the testing.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>