Update raw_connect to simplify validation from UI
The UI repo has changed the way endpoint validation occurs, causing a
problem withinin the angular controller when creating the arguments for
`raw_connect`. Previously, the `auth_url` was required as an argument
which mandated the need to expose the `auth_url` from the Nuage manager
so that it could be used in the UI controller.

Instead, this patch replaces `auth_url` with the specific arguments that
are needed to build the `auth_url` so that the UI can directly pass
those values into the manager.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>