Enable Nuage network manager
Prior to this patch, the UI could not show the option to create new
network managers because these are typically created along with cloud
manager. Nuage is a dedicated SDN requiring ability to add the provider
manually.

The UI depends on the `product.nuage` setting to enable the network
providers which this patch adds directly into the Nuage.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>