Batch collecting of Nuage events with qpid sleeping
With this commit we force periodic interrupts to the EventCatcher
thread because the gem for accessing the AMQP queue seems to consume
all EventCatcher thread resources otherwise. By default we now interrupt
the qpid gem every 15 seconds and let it sleep for 15 seconds. While
the qpid thread is sleeping, the EventProcessing thread has enough
resources to quickly process events that were captured so far.

It is not clear at the moment why resources are not distributed equally
among the threads - could be a leak in amqp_proton or in MIQ. Below please
find description of old and new event catcher behavior:

--- OLD BEHAVIOR ---
Events were reported immediately by qpid_proton gem, but they didn't get
processed by ManageIQ::Providers::Nuage::NetworkManager::EventParser for quite
a while then. Since events were being continously collected in the queue and
EventParser didn't have resources to process them the delay was increasing.
Soon after MIQ boot the delay between receiveing event and processing event
was as huge as 10 min.

--- NEW BEHAVIOR ---
Events are reported while the qpid thread is running and they are all processed
instantly when the qpid thread is put to sleep. So each event gets processed
in about 15 seconds, which is acceptable.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>