Improve the performance of event catcher
Nuage event catcher relies on Qpid Proton to get the events from the
AMQP queue. Due to synchronization issues, this resulted in massive
delays when processing events.

This patch tries to overcome this by adding a short sleep cycle to allow
the stream (Qpid) and event catcher to work together on the events.

It also resolves an issue with the fact the Nuage events are sometimes
sent with the same timestamp for certain event types, even though the
actual payload differs. As there is no ID that could be used as an EMS
reference, we alter the timestamp slightly based on the current time.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>