Upgrade qpid_proton related stuff to support v0.19.0
With this commit we adjust whatever needed to support qpid_proton
version 0.19.0 instead of 0.18.1. Even if only a "minor" number has
increased, the API is broken hence this commit. Most notable
differences in qpid_proton API are:

- username/password can no longer be provided as part of URL, but separately
- class Qpid::Proton::Handler::MessagingHandler has been replaced with Qpid::Proton::Handler::MessagingHandler
- some handler callbacks have been renamed (on_start -> on_container_start etc.)
- some other functions have been renamed and signature has been updated
- :urls option is no longer supported (connection to fallback endpoint had to be implemented on MIQ level)
- timeout is no hardcoded to 60sec

The reason to switch to qpid_proton 0.19.0 is that we measured huge
performance issues on 0.18.1 that seem to be fixed in 0.19.0.

Fixes https://github.com/ManageIQ/manageiq-providers-nuage/issues/47

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>