Assign :ems_ref to the event
With this commit we make sure that events are always created with
unique ems_ref. For most of them we just grab value from the event
hash. We make up a new unique value when ID is not present in event
hash to make is unique at least from this point on.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>