Restart EventCatcher everytime we update provider
Prior this commit EventCatcher wasn't restarted when AMQP endpoints
or fallback endpoints or port were updated, which is annoying. User
had to either manually `sudo kill -9 PID` the old EventCatcher or
restart the whole server.

With this commit we make sure that stop_event_monitor is called
everytime when Nuage provider is updated.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>