Don't run EventCatcher when user opts-in for "None"
Nuage GUI offers option "None" for eventing (as opposed to "AMQP").
Turns out that EventCatcher restarts itself like crazy in case
"None" is chosen because it tries to connect to AMQP anyway even if
it lacky both endpoint information and credentials.

With this commit we make sure that EventCatcher is only run for those
Nuage EMSes that have option "AMQP" actually chosen.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>