Silently reconnect qpid in case of heartbeat timeout
We observed that Nuage server often "forgets" to send AMQP heartbeat
for some reason which results in Nuage provider's event catcher
reporting the error in evm.log and reconnecting to the next fallback
url.

With this commit we capture such type of exception and reconnect
silently to the same fallback url to prevent event catcher from looping
fallback urls for this error and also to keep evm.log clear.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>