Move qpid_proton dependency from core repo to nuage repo
With this commit we introduce qpid_proton dependency directly here
on Nuage provider. It was originally put to the core repo because
it was only installable from github repository. This is no longer the
case as gem is available in RubyGems repository:

```
https://rubygems.org/gems/qpid_proton
```

We want to be able to shift gem version when Nuage provider
integration proceeds without affecting the core repo therefore
it's better for it to reside here.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>