Narrow down focus of targeted refresh
When event reports modification of CloudSubnet current targeted refresh
performs way too much work: it refreshes not only this specific CloudSubnet,
but also all related entities - CloudTenant, NetworkRouter, SecurityGroup etc.
This results in 40 API calls instead of one!

Rationale behind was that we wanted to make sure all the relevant entities get
updated as well BUT it turns out Nuage yields events for every piece of change
anyway so we can safely assume that targeted refresh will handle them.

With this commit we simplify targeted refresh to not infer any related entities
and rather just focus on updating the specific one that it was invoked for.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>