Targeted refresh for L2 and L3 CloudSubnets
With this commit we adjust targeted refresh to works for both kind
of subnets - CloudSubnet::L2 and CloudSubnet::L3. When we recieve
event we need to pass additional information to target in order to
later know what type of subnet are we updating. We need to know this
because different API calls need to be used for L2 and L3 subnets...

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>