Reduce memoization in targeted refresh
We're doing a lot of unneccessary caching in targeted collector.
Not anymore - with this commit we simplify the code and spare memory.
Rspecs are left intact and there are no errors about missing http interactions
which confirms that caching was never used - because each entity is
fetched and parsed only once anyway.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>