Overcome missing event glitch of Nuage server
With this commit we overcome a bug on Nuage side that
upon instantiating router template, no events are emitted
for subnets.

In other words: when we instantiate template that also contains
subnets, MIQ is not explicitly informed that subnets were created,
because events are missing. So we play smart: when "router create"
event is triggered, we infer all its subnets as well.

NOTE: glitch only appears when we CREATE router. Updating/deleting
router properly emits events so we're only fixing this specific case.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>