Overcome another missing event glitch of Nuage server
With this commit we overcome a bug on Nuage side that upon
deletion of NetworkRouter, corresponding SecurityGroups are
also being deleted, but MIQ is not informed about it because
no event is emitted.

We overcome this by inferring all related SecurityGroups when
"router was deleted" event occurs. Unfortunately, MIQ does not
yet modell relation "security group belongs to network router"
so we have to perform check for all SecurityGroups in the
CloudTenant.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>