Remove CloudSubnet when corresponding subnet template is deleted
With this commit we tackle Nuage glitch that event is not triggered.
So when subnet template is deleted, also all the subnets that are
connected to it are deleted, but no explicit event is triggered hence
MIQ does not remove them from VMDB.

Solution: we now capture "subnet template was deleted" event and
manually remove all subnets linked to subnet template.

To get things even better: MIQ is currently not inventoring subnet templates
nor supports anything like this in its schema. Therefore we abuse
`cloud_subnet.extra_attributes` to store subnet template ID in it so
that we can later know what subnets are linked to which subnet template.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>