Targeted refresh: remove subnet when zone is removed
There is a known Nuage glitch that some events are not being
emitted. With this commit we tackle the one that when zone or
zone template are deleted, also all the subnets in it are deleted,
but MIQ has no idea about it. So we need to play smart: when
recieveing either of

```
nuage_zonetemplate_delete
nuage_zone_delete
```

events, we manually mark also all corresponding subnets for
deletion.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>