Cross-connect NetworkPorts to VMs
With this commit we extend inventoring of `NetworkPort::Vm` model
to now connect it to a VM. We assume that VM is in VMDB at the time
refresh takes place. It will be connected regardless what provider
the VM belongs to.

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1619544

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>