Unittest .travis.yml to ensure excludes match the matrix
When we're shifting supported ruby versions we often forget
to update the "excludes" block which in turn means we're then
re-running javascript tests more than we need to. Now, there is
no big harm in re-running the test more than once, but still.

With this commit we unit test that excludes block matches
values from the matrix.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>